requirements.txt
boto3==1.34.10
python-dotenv==1.0.0
